Added TR069 - 'network breadboarding system'
This repository contains everything to create and configure a virtual TR-069 test network
utilising docker to run openACS or genieACS, XMPP and STUN in combination with different
private and public networks protected by gateways and firewalls to be able to connect
and test your DUT.

The intended use case here at devolo is to test some TR-069 aspects in combination with
our products. For productive uses the credentials and the setup of this 'network
breadboarding system' is highly insecure.

Nethertheless the 'network breadboarding system' is way more than that, you are able to easyly
extend your virtual setup by plugging networks and systems containing real world applications
together like ICs on a breadboard.