Replaces static mocks by create_mocks.sh
You can define a profile with --profile=<NAME>. The profile, if not
being build in, will be represented by a directory. You can copy
the directory to a new one, change the content of the files to the
desired ones and use the new directory as a new profile.
A not existing profile will be created as directory and filled
with default values.
You can use as may profiles you like, they are applied in the
specified order and will overwrite settings.

  Build in profiles:
     default -> Speedport Smart 3
                    * DEVICE_SpeedportSmart3
                    * TR181
                    * DATASET_1

     DEVICE_SpeedportSmart3 \
     DEVICE_Fritz7590        }-> mocking this device
     DEVICE_Fritz7490       /
     TR181   -> responding to GetParameterValues for
                Device.DSL.Line.1.TestParams. with 200 and data
     TR098   -> responding to X_GENERIC_GetVDSLInfo with
                HTTP 200 and data
     200 -> will respond with HTTP 200 and data
     500 -> will respond with HTTP 500 error 401 Invalid Action
     DATASET_1 -> one data set from speedport
     DATASET_2 -> another data set from fritzbox

The mock can then be found using SSDP and will serve the data from
the selected profile directory.
Also the IP address to be announced can be set by --IP=<IP>,
default is 127.0.0.1